% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb.R
\name{sb_children}
\alias{sb_children}
\title{List child items of a ScienceBase item}
\usage{
sb_children(parent_id, max = 20, offset = 0)
}
\arguments{
\item{parent_id}{Parent item ID}

\item{max}{Number of results (default 20, max 100)}

\item{offset}{Pagination offset (default 0)}
}
\value{
tibble: id, title, summary, hasChildren, dateCreated, lastUpdated, tags
}
\description{
List child items of a ScienceBase item
}
