% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb.R
\name{sb_item}
\alias{sb_item}
\title{Get full details for a ScienceBase item}
\usage{
sb_item(item_id)
}
\arguments{
\item{item_id}{ScienceBase item ID}
}
\value{
tibble with one row: id, title, summary, hasChildren, dateCreated,
lastUpdated, tags, contacts, files
}
\description{
Get full details for a ScienceBase item
}
